Add Jared Roth's Blog
Just noticed that I somehow screwed up the previous pull request for this and don't see it here, open or closed. So here it is again.
I did not bother with any customization beyond content because I have been out of town this week and felt like that was the safest area to cut back on some time.
I do plan on customizing it over time.