Added Alireza Andersen link
Sorry Andrew I hope this corrects the merge conflict, I really don't know why or how a merge conflict happened. 

Biggest struggle was feature testing password protection and figure out a way to change html link colors. 

The controller and Model have the methods for the authorization under delete. It is the feature test_user_can_delete_robot. I originally made the test pass by asserting that the path changes. But I realize that even though it passes, it is not a strong test. So I tried to assert a change in count with the deletion of robots. The issue I am running into is the authorization of access during the test. I followed the link(to the best of my ability). Any guidance and feed back is great appreciated.  Have a great night!